BasicCombat
- Separate input control and move control
