+++ Assets/Scenes/Sandbox.unity
@@ -1796,3 +1796,60 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 1f80192e74fc2d74b9da3ea5ee9e7d90, type: 3}
+--- !u!1001 &8415911487619304399
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 5815418572369954607, guid: 002ddb70de8718041b1c007093d10687, type: 3}
+      propertyPath: m_RootOrder
+      value: 4
+      objectReference: {fileID: 0}
+    - target: {fileID: 5815418572369954607, guid: 002ddb70de8718041b1c007093d10687, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5815418572369954607, guid: 002ddb70de8718041b1c007093d10687, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.46
+      objectReference: {fileID: 0}
+    - target: {fileID: 5815418572369954607, guid: 002ddb70de8718041b1c007093d10687, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: -4.38
+      objectReference: {fileID: 0}
+    - target: {fileID: 5815418572369954607, guid: 002ddb70de8718041b1c007093d10687, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5815418572369954607, guid: 002ddb70de8718041b1c007093d10687, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5815418572369954607, guid: 002ddb70de8718041b1c007093d10687, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5815418572369954607, guid: 002ddb70de8718041b1c007093d10687, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5815418572369954607, guid: 002ddb70de8718041b1c007093d10687, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5815418572369954607, guid: 002ddb70de8718041b1c007093d10687, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5815418572369954607, guid: 002ddb70de8718041b1c007093d10687, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8790387743193397837, guid: 002ddb70de8718041b1c007093d10687, type: 3}
+      propertyPath: m_Name
+      value: Enemy
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 002ddb70de8718041b1c007093d10687, type: 3}
