+++ Assets/Scripts/Control.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2db268918dfeb3f4199b9fd0cbcf258d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
