+++ Assets/Scripts/Movement/Mover.cs
@@ -1,14 +1,8 @@
-using System.Collections;
-using UnityEditor;
 using UnityEngine;
 using UnityEngine.AI;
 
 public class Mover : MonoBehaviour {
     void Update() {
-        if (Input.GetMouseButton(0)) {
-            MoveToCursor();
-        }
-
         UpdateAnimator();
     }
 
@@ -19,12 +13,7 @@ public class Mover : MonoBehaviour {
         GetComponent<Animator>().SetFloat("fowardSpeed", speed);
     }
 
-    private void MoveToCursor() {
-        Ray ray = Camera.main.ScreenPointToRay(Input.mousePosition);
-        RaycastHit hit;
-        bool hasHit = Physics.Raycast(ray, out hit);
-        if (hasHit) {
-            GetComponent<NavMeshAgent>().destination = hit.point;
-        }
+    public void MoveTo(Vector3 destination) {
+        GetComponent<NavMeshAgent>().destination = destination;
     }
 }
